pyproject.toml
[project]
name = "voxel"
version = "0.1.0"
description = "Pocket AI companion device — animated cube mascot with voice interaction"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11,<3.14"
dependencies = [
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "paramiko>=3.4.0",
    "edge-tts>=6.1.0",
    "websockets>=12.0",
    "rich>=13.0",
    "watchfiles>=1.0.0",
    "qrcode>=7.0",
]

[project.optional-dependencies]
pygame = [
    "pygame>=2.5.0",
]

pi = [
    "spidev>=3.6",
    "RPi.GPIO>=0.7.1",
]

dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.2",
    "sounddevice>=0.4.6",
    "miniaudio>=1.50",
]

mcp = [
    "websockets>=12.0",  # already a core dep, but explicit for mcp standalone
]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 10

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["cli", "core", "config", "display", "hw", "mcp", "shared", "states"]

[project.scripts]
voxel = "cli:main"
dev = "display.service:main"
dev-watch = "display.service:main_watch"
mcp-server = "mcp.server:main"
